removed RotationPresets presets and just made it dynamic changing in GetRotations
