+++ src/main/java/apple/trees/tree/trunk/creation/utils/GetRotations.java
@@ -7,6 +7,7 @@ import org.bukkit.plugin.java.JavaPlugin;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.Random;
 
 public class GetRotations {
@@ -38,8 +39,44 @@ public class GetRotations {
     }
 
     private static ArrayList<Vec3d> rotationBaseFromDomain(int branchesToBuild) {
-        ArrayList<Vec3d> rotations = RotationPresets.branchesFromDomain(branchesToBuild);
+        ArrayList<Vec3d> rotations = branchesFromDomain(branchesToBuild);
         return rotations;
+    }
+    protected static ArrayList<Vec3d> branchesFromDomain(int numOfBranches) {
+        ArrayList<Vec3d> polars = new ArrayList<>();
+        for (int i = 0; i < numOfBranches; i++) {
+            polars.add(new Vec3d());
+        }
+
+        //fill arr with the different vals that the components could be
+        ArrayList<Double> arr = new ArrayList<>();
+        double increment = 720.0 / (numOfBranches - 1);
+        for (double i = -360; i <= 360; i += increment) {
+            arr.add(i);
+        }
+        int size = polars.size();
+        if (size != arr.size()) {
+            System.err.println("somefin went wrong in making the increments");
+            for (Double i : arr) {
+                System.err.print(i + " ");
+            }
+            System.err.println();
+        }
+        // for each component, add the correct component to each branch
+        for (int i = 0; i < size; i++) {
+            Collections.shuffle(arr);
+            polars.get(i).x = arr.get(i);
+        }
+        for (int i = 0; i < size; i++) {
+            Collections.shuffle(arr);
+            polars.get(i).y = arr.get(i);
+        }
+        for (int i = 0; i < size; i++) {
+            Collections.shuffle(arr);
+            polars.get(i).z = arr.get(i);
+        }
+
+        return polars;
     }
 
 }
